requirements.txt
flask
pydantic
gunicorn
opentelemetry-distro 
opentelemetry-exporter-otlp 
opentelemetry-instrumentation-flask